\name{cancerGI-package}
\alias{cancerGI-package}
\alias{cancerGI}
\docType{package}
\title{
Statistical Analyses of Genetic Interactions in Cancer
}
\description{
This package contains statistical analyses performed in Wang, Fu, et al. (2014) for the genomic data from the TCGA breast cancer patients, as well as for the RNAi knockdown data.
}
\details{
\tabular{ll}{
Package: \tab cancerGI\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-19\cr
License: \tab What license is it under?\cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Audrey Qiuyan Fu, Xiaoyue Wang

Maintainer: Audrey Fu <audreyqyfu@gmail.com>
}
\references{
Wang, X., Fu, A. Q., McNerney, M. and White, K. P. (2014). Widespread genetic epistasis among breast cancer genes. Nature Communications. 5 4828. doi: 10.1038/ncomms5828
}
\keyword{ package }
\seealso{
\code{\link{RNAi}}
\code{\link{testMutationalPatternAll.wrapper}}
\code{\link{computeSurvivalPValueForGenePairSet.output}}
\code{\link{computeSmallWorldness}}
}
