\name{tested_pairs}
\alias{tested_pairs}
\docType{data}
\title{Gene pairs tested in the double knockdown assay.
}
\description{
It contains two columns of gene names.
}
%%\usage{data("tested_pairs")}
\format{
  A data frame with 1508 observations on the following 2 variables.
  \describe{
    \item{\code{V1}}{a character vector}
    \item{\code{V2}}{a character vector}
  }
}
\references{
Wang, X., Fu, A. Q., McNerney, M. and White, K. P. (2014). Widespread genetic epistasis among breast cancer genes. Nature Communications. 5 4828. doi: 10.1038/ncomms5828
}
\examples{
data(tested_pairs)
## see documentation for dataset \code{\link{RNAi}}
}
\keyword{datasets}
